analytics/requirements.txt
neo4j>=5.0
python-igraph>=0.11
leidenalg>=0.10
